+++ .env.example
@@ -29,6 +29,7 @@ OPENAI_ORG_ID=your_openai_org_id
 FIREWORKS_API_KEY=your_fireworks_api_key
 STABILITY_API_KEY=your_stability_api_key
 BRIA_AI_API_KEY=your_bria_ai_api_key
+DID_API_KEY=your_did_api_key
 
 NEXT_PUBLIC_COOKIE_NAME=generateAuthToken
 
@@ -41,3 +42,5 @@ NEXT_PUBLIC_CREDITS_PER_STABLE_DIFFUSION_XL_IMAGE=4
 NEXT_PUBLIC_CREDITS_PER_STABILITY_SD3_TURBO_IMAGE=4
 NEXT_PUBLIC_CREDITS_PER_PLAYGROUND_V2_IMAGE=4
 NEXT_PUBLIC_CREDITS_PER_BRIA_IMAGE=4
+NEXT_PUBLIC_CREDITS_PER_PLAYGROUND_V2_5_IMAGE=4
+NEXT_PUBLIC_CREDITS_PER_D_ID=50
