+++ src/components/GenerateImage.tsx
@@ -21,7 +21,7 @@ import { getLightingFromLabel, lightings } from "@/constants/lightings";
 import { useSearchParams } from "next/navigation";
 import CreatableSelect from "react-select/creatable";
 import { suggestTags } from "@/actions/suggestTags";
-import { Image as ImageIcon, Mic, StopCircle, XCircle } from "lucide-react";
+import { Image as ImageIcon, Mic, StopCircle, VideoIcon, XCircle } from "lucide-react";
 import { imageCategories } from "@/constants/imageCategories";
 
 interface SpeechRecognitionEvent extends Event {
@@ -40,6 +40,7 @@ interface SpeechRecognitionErrorEvent extends Event {
 
 // Import the server action
 import { generateImage } from "@/actions/generateImage";
+import { audios } from "@/constants/audios";
 
 export default function GenerateImage() {
   const uid = useAuthStore((s) => s.uid);
@@ -52,20 +53,26 @@ export default function GenerateImage() {
   const tagsSearchParam = searchterm.get("tags")?.split(",");
   const imageReferenceSearchParam = searchterm.get("imageReference");
   const imageCategorySearchParam = searchterm.get("imageCategory");
+  const videoModelSearchParam = searchterm.get("videoModel");
+  const audioSearchParam = searchterm.get("audio");
+  const scriptPromptSearchParam = searchterm.get("scriptPrompt");
   const fireworksAPIKey = useProfileStore((s) => s.profile.fireworks_api_key);
   const openAPIKey = useProfileStore((s) => s.profile.openai_api_key);
   const stabilityAPIKey = useProfileStore((s) => s.profile.stability_api_key);
+  const didApiKey = useProfileStore((s) => s.profile.did_api_key);
   const useCredits = useProfileStore((s) => s.profile.useCredits);
   const credits = useProfileStore((s) => s.profile.credits);
   const minusCredits = useProfileStore((state) => state.minusCredits);
   const [imagePrompt, setImagePrompt] = useState<string>(
     freestyleSearchParam || ""
   );
+  const [scriptPrompt, setScriptPrompt] = useState<string>(scriptPromptSearchParam || "")
 
   const [imageStyle, setImageStyle] = useState<string>(styleSearchParam || "");
   const [model, setModel] = useState<model>(
     (modelSearchParam as model) || "playground-v2"
   );
+  const [videoModel, setVideoModel] = useState<model>(videoModelSearchParam as model || "d-id");
   const [colorScheme, setColorScheme] = useState<string>(
     colorSearchParam || "None"
   );
@@ -78,16 +85,19 @@ export default function GenerateImage() {
   const [tagInputValue, settagInputValue] = useState(
     tagsSearchParam
       ? tagsSearchParam.map((str) => {
-          return { label: str, value: str };
-        })
+        return { label: str, value: str };
+      })
       : []
   );
   const [suggestedTags, setSuggestedTags] = useState<string[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
   const [generatedImage, setGeneratedImage] = useState<string>("");
+  const [generatedVideo, setGeneratedVideo] = useState<string>("");
   const [isRecording, setIsRecording] = useState<boolean>(false);
   const [uploadedImage, setUploadedImage] = useState<File | null>(null);
   const [selectedCategory, setSelectedCategory] = useState<string>(imageCategorySearchParam || "");
+  const [mode, setMode] = useState(scriptPromptSearchParam ? "video" : "image");
+  const [audio, setAudio] = useState<string>(audioSearchParam || "Male");
 
   const [promptData, setPromptData] = useState<PromptDataType>({
     style: "",
@@ -99,6 +109,25 @@ export default function GenerateImage() {
     lighting: getLightingFromLabel(lighting) || lightings[0].value,
     tags: tags
   });
+
+  const [isPromptValid, setIsPromptValid] = useState<boolean>(false);
+  const [isScriptPromptValid, setIsScriptPromptValid] = useState<boolean>(true);
+  const [isModelValid, setIsModelValid] = useState<boolean>(true);
+  const [isVideoModelValid, setIsVideoModelValid] = useState<boolean>(true);
+  const [isAudioValid, setIsAudioValid] = useState<boolean>(true);
+
+  useEffect(() => {
+    setIsPromptValid(!!imagePrompt.trim());
+    setIsModelValid(!!model);
+    if (mode === 'video') {
+      setIsScriptPromptValid(!!scriptPrompt.trim());
+      setIsVideoModelValid(!!videoModel);
+      setIsAudioValid(!!audio);
+    } else {
+      setIsScriptPromptValid(true);
+      setIsVideoModelValid(true);
+    }
+  }, [imagePrompt, scriptPrompt, model, videoModel, mode]);
 
   const colorLabels = colors.map(
     (color: { value: string; label: string }) => color.label
@@ -178,27 +207,40 @@ export default function GenerateImage() {
   async function saveHistory(
     promptData: PromptDataType,
     prompt: string,
-    downloadUrl: string
+    downloadUrl: string,
+    options: { videoDownloadUrl?: string; audio?: string, videoModel?: string, scriptPrompt?: string } = {}
   ) {
     if (!uid) return;
+
     const coll = collection(db, "profiles", uid, "covers");
     const docRef = doc(coll);
-    const p: PromptDataType = {
+
+    const finalPromptData: PromptDataType = {
       ...promptData,
-      downloadUrl: downloadUrl,
-      model: model,
-      prompt: prompt,
+      downloadUrl,
+      model,
+      prompt,
       id: docRef.id,
       timestamp: Timestamp.now(),
       tags,
-      imageCategory: selectedCategory
+      imageCategory: selectedCategory,
+      videoDownloadUrl: options.videoDownloadUrl || "",
+      audio: options.audio || "",
+      videoModel: options.videoModel || "",
+      scriptPrompt: options.scriptPrompt
     };
-    setPromptData(p);
-    await setDoc(docRef, p);
+
+    setPromptData(finalPromptData);
+    await setDoc(docRef, finalPromptData);
   }
 
   const handleGenerateSDXL = async (e: React.FormEvent<HTMLButtonElement>) => {
     e.preventDefault();
+
+    if (!isPromptValid || !isModelValid || (mode === 'video' && (!isScriptPromptValid || !isVideoModelValid))) {
+      toast.error("Please fill in all required fields.");
+      return;
+    }
 
     try {
       setLoading(true);
@@ -217,30 +259,42 @@ export default function GenerateImage() {
       formData.append("openAPIKey", openAPIKey);
       formData.append("fireworksAPIKey", fireworksAPIKey);
       formData.append("stabilityAPIKey", stabilityAPIKey);
+      formData.append("didAPIKey", didApiKey);
       formData.append("useCredits", useCredits.toString());
       formData.append("credits", credits.toString());
       formData.append("model", model);
       if (uploadedImage) {
         formData.append("imageField", uploadedImage);
       }
+      if (scriptPrompt) {
+        formData.append("scriptPrompt", scriptPrompt);
+        formData.append("videoModel", videoModel);
+        formData.append("audio", audio);
+      }
 
       // Call the server action instead of the API route
       const result = await generateImage(formData);
 
       // Updated error handling
-      if (!result || !result.imageUrl) {
-        toast.error("Failed to generate image.");
-        throw new Error("Failed to generate image.");
+      if (!result || (!result.videoUrl && !result.imageUrl)) {
+        toast.error(`Failed to generate image/video: ${result.error}`);
+        throw new Error("Failed to generate image/video.");
       }
 
-      const downloadURL = result.imageUrl;
+      const downloadURL = result.imageUrl || "";
+      const videoDownloadURL = result.videoUrl || "";
       const imageReference = result.imageReference || "";
 
       if (useCredits) {
         await minusCredits(creditsToMinus(model));
+
+        if (videoDownloadURL) {
+          await minusCredits(creditsToMinus(videoModel))
+        }
       }
 
       setGeneratedImage(downloadURL);
+      setGeneratedVideo(videoDownloadURL);
 
       await saveHistory(
         {
@@ -248,15 +302,18 @@ export default function GenerateImage() {
           freestyle: imagePrompt,
           style: imageStyle,
           downloadUrl: downloadURL,
-          model: model,
+          videoDownloadUrl: videoDownloadURL,
+          model,
           prompt,
           lighting: getLightingFromLabel(lighting) || lightings[0].value,
           colorScheme: getColorFromLabel(colorScheme) || colors[0].value,
           imageReference,
-          imageCategory: selectedCategory
+          imageCategory: selectedCategory,
+          audio: audio
         },
         prompt,
-        downloadURL
+        downloadURL,
+        { videoDownloadUrl: videoDownloadURL, audio: audio, videoModel: videoModel, scriptPrompt: scriptPrompt }
       );
     } catch (error: unknown) {
       if (error instanceof Error) {
@@ -275,12 +332,14 @@ export default function GenerateImage() {
         <div className="relative">
           <TextareaAutosize
             autoFocus
-            minRows={4}
+            minRows={mode == "image" ? 4 : 3}
             value={imagePrompt || ""}
-            placeholder="Describe an image or use voice input"
+            placeholder={mode == "image" ? "Describe an image or use voice input" : "Describe the avatar or use voice input"}
             onChange={(e) => {
               setImagePrompt(e.target.value);
-              handleTagSuggestions(e.target.value);
+              if (mode === "image") {
+                handleTagSuggestions(e.target.value);
+              }
             }}
             className="border-2 text-xl border-blue-500 bg-blue-100 rounded-md px-3 py-2 w-full"
           />
@@ -357,6 +416,18 @@ export default function GenerateImage() {
           </div>
         )}
 
+        {mode === "video" && <TextareaAutosize
+          autoFocus
+          minRows={4}
+          value={scriptPrompt || ""}
+          placeholder="Write the script here"
+          onChange={(e) => {
+            setScriptPrompt(e.target.value);
+            handleTagSuggestions(e.target.value);
+          }}
+          className="border-2 text-xl border-blue-500 bg-blue-100 rounded-md px-3 py-2 w-full"
+        />}
+
         <div>
           <div>Artistic Style (optional)</div>
           <Select
@@ -369,8 +440,9 @@ export default function GenerateImage() {
             defaultInputValue={styleSearchParam || ""}
           />
         </div>
+
         <div>
-          <div>Use</div>
+          <div>{mode === "image" ? "Use" : "Image Model (Avatar Creation)"}</div>
           <Select
             isClearable={true}
             isSearchable={true}
@@ -381,10 +453,45 @@ export default function GenerateImage() {
             defaultValue={findModelByValue(
               (modelSearchParam as model) || "playground-v2"
             )}
-            options={models}
+            options={models.filter((m) => m.type === "image")}
             styles={selectStyles}
           />
         </div>
+
+        {mode === "video" &&
+          <div>
+            <div>Video Model</div>
+            <Select
+              isClearable={true}
+              isSearchable={true}
+              name="videoModel"
+              onChange={(v) =>
+                setVideoModel(v ? (v as SelectModel).value : "d-id")
+              }
+              defaultValue={findModelByValue("d-id")}
+              options={models.filter((m) => m.type === "video")}
+              styles={selectStyles}
+            />
+          </div>
+        }
+
+        {mode === "video" && <div>
+          <div>Audio</div>
+          <Select
+            isClearable={true}
+            isSearchable={true}
+            name="audio"
+            onChange={(v) => setAudio(v ? v.value : "Male")}
+            options={audios.map((audio) => ({
+              id: audio.id,
+              label: audio.label,
+              value: audio.value,
+            }))}
+            defaultInputValue={"Male"}
+            styles={selectStyles}
+            placeholder="Select audio"
+          />
+        </div>}
 
         <div>
           <div>Image Category (optional)</div>
@@ -424,9 +531,8 @@ export default function GenerateImage() {
             {colorLabels.map((option) => (
               <div
                 key={option}
-                className={`cursor-pointer flex items-center space-x-1 p-2 rounded-md ${
-                  colorScheme === option ? "bg-gray-200" : ""
-                }`}
+                className={`cursor-pointer flex items-center space-x-1 p-2 rounded-md ${colorScheme === option ? "bg-gray-200" : ""
+                  }`}
                 onClick={() => setColorScheme(option)}
                 title={option}
               >
@@ -442,9 +548,8 @@ export default function GenerateImage() {
             {lightingLabels.map((option) => (
               <div
                 key={option}
-                className={`cursor-pointer flex items-center space-x-1 p-2 rounded-md ${
-                  lighting === option ? "bg-gray-200" : ""
-                }`}
+                className={`cursor-pointer flex items-center space-x-1 p-2 rounded-md ${lighting === option ? "bg-gray-200" : ""
+                  }`}
                 onClick={() => setLighting(option)}
                 title={option}
               >
@@ -452,6 +557,26 @@ export default function GenerateImage() {
               </div>
             ))}
           </div>
+        </div>
+
+        <div className="flex space-x-4 mb-4">
+          <button
+            className={`w-10 h-10 flex items-center justify-center rounded-full 
+              ${mode === 'image' ? 'bg-blue-600' : 'bg-gray-300'} text-white`}
+            onClick={() => { setMode('image'); setScriptPrompt(""); }}
+            title="Image Mode"
+          >
+            <ImageIcon size={20} />
+          </button>
+
+          <button
+            className={`w-10 h-10 flex items-center justify-center rounded-full 
+              ${mode === 'video' ? 'bg-blue-600' : 'bg-gray-300'} text-white`}
+            onClick={() => setMode('video')}
+            title="Video Mode"
+          >
+            <VideoIcon size={20} />
+          </button>
         </div>
 
         <button
@@ -469,12 +594,18 @@ export default function GenerateImage() {
             handleGenerateSDXL(e);
           }}
         >
-          {loading ? <PulseLoader color="#fff" size={12} /> : "Create Image"}
+          {loading ? <PulseLoader color="#fff" size={12} /> : mode === "image" ? "Create Image" : 'Create Video'}
         </button>
       </div>
 
-      <div className="w-full max-w-2xl mt-6">
-        {generatedImage ? (
+      <div className="w-full max-w-2xl mt-6 flex justify-center">
+        {generatedVideo ? (
+          <video
+            className="object-cover rounded-md"
+            src={generatedVideo}
+            controls
+          />
+        ) : generatedImage ? (
           <img
             className="object-cover rounded-md"
             src={generatedImage}
@@ -482,7 +613,7 @@ export default function GenerateImage() {
           />
         ) : (
           <div className="text-gray-500 text-center">
-            No image generated yet.
+            No video or image generated yet.
           </div>
         )}
       </div>
