+++ src/components/ProfileComponent.tsx
@@ -12,6 +12,7 @@ export default function ProfileComponent() {
   const [openaiApiKey, setOpenaiApiKey] = useState(profile.openai_api_key);
   const [stabilityAPIKey, setStabilityAPIKey] = useState(profile.stability_api_key);
   const [briaApiKey, setBriaApiKey] = useState(profile.bria_api_key);
+  const [didApiKey, setdidApiKey] = useState(profile.did_api_key);
 
   const [useCredits, setUseCredits] = useState(profile.useCredits);
   const [showCreditsSection, setShowCreditsSection] = useState(true);
@@ -21,6 +22,7 @@ export default function ProfileComponent() {
     setOpenaiApiKey(profile.openai_api_key);
     setStabilityAPIKey(profile.stability_api_key);
     setBriaApiKey(profile.bria_api_key);
+    setdidApiKey(profile.did_api_key);
 
     setShowCreditsSection(!isIOSReactNativeWebView());
   }, [
@@ -28,6 +30,7 @@ export default function ProfileComponent() {
     profile.openai_api_key,
     profile.stability_api_key,
     profile.bria_api_key,
+    profile.did_api_key
   ]);
 
   const handleApiKeyChange = async () => {
@@ -35,7 +38,8 @@ export default function ProfileComponent() {
       fireworksApiKey !== profile.fireworks_api_key ||
       openaiApiKey !== profile.openai_api_key ||
       stabilityAPIKey !== profile.stability_api_key ||
-      briaApiKey !== profile.bria_api_key
+      briaApiKey !== profile.bria_api_key ||
+      didApiKey !== profile.did_api_key
     ) {
       try {
         await updateProfile({
@@ -43,6 +47,7 @@ export default function ProfileComponent() {
           openai_api_key: openaiApiKey,
           stability_api_key: stabilityAPIKey,
           bria_api_key: briaApiKey,
+          did_api_key: didApiKey
         });
         console.log("API keys updated successfully!");
       } catch (error) {
@@ -67,7 +72,7 @@ export default function ProfileComponent() {
     [showCreditsSection]
   );
 
-  const areApiKeysAvailable = fireworksApiKey && openaiApiKey && stabilityAPIKey && briaApiKey;
+  const areApiKeysAvailable = fireworksApiKey && openaiApiKey && stabilityAPIKey && briaApiKey && didApiKey;
 
   return (
     <div className="flex flex-col gap-4">
@@ -135,13 +140,26 @@ export default function ProfileComponent() {
           placeholder="Enter your Bria API Key"
         />
 
+        <label htmlFor="d-id-api-key" className="text-sm font-medium">
+          D-ID API Key:
+        </label>
+        <input
+          type="text"
+          id="d-id-api-key"
+          value={didApiKey}
+          onChange={(e) => setdidApiKey(e.target.value)}
+          className="border border-gray-300 rounded-md px-3 py-2 h-10"
+          placeholder="Enter your D-ID API Key"
+        />
+
         <button
           onClick={handleApiKeyChange}
           disabled={
             fireworksApiKey === profile.fireworks_api_key &&
             openaiApiKey === profile.openai_api_key &&
             stabilityAPIKey === profile.stability_api_key &&
-            briaApiKey === profile.bria_api_key
+            briaApiKey === profile.bria_api_key &&
+            didApiKey === profile.did_api_key
           }
           className="bg-blue-500 text-white px-3 py-2 rounded-md hover:opacity-50 disabled:opacity-50"
         >
