+++ src/constants/audios.ts
@@ -1,0 +1,12 @@
+export const audios = [
+    {
+        id: 1,
+        value: "male",
+        label: "Male",
+    },
+    {
+        id: 2,
+        value: "female",
+        label: "Female",
+    },
+]
