+++ src/utils/credits.ts
@@ -13,8 +13,12 @@ export const creditsToMinus = (model: model): number => {
         return parseEnvVarToNumber(process.env.NEXT_PUBLIC_CREDITS_PER_STABILITY_SD3_TURBO_IMAGE, 4);
     } else if (model == 'playground-v2') {
         return parseEnvVarToNumber(process.env.NEXT_PUBLIC_CREDITS_PER_PLAYGROUND_V2_IMAGE, 4);
+    } else if (model == "playground-v2-5") {
+        return parseEnvVarToNumber(process.env.NEXT_PUBLIC_CREDITS_PER_PLAYGROUND_V2_5_IMAGE, 4);
     } else if (model == 'bria.ai') {
         return parseEnvVarToNumber(process.env.NEXT_PUBLIC_CREDITS_PER_BRIA_IMAGE, 4);
+    } else if (model == 'd-id') {
+        return parseEnvVarToNumber(process.env.NEXT_PUBLIC_CREDITS_PER_D_ID, 50);
     }
 
     return 4;
