+++ src/zustand/useProfileStore.ts
@@ -14,6 +14,7 @@ export interface ProfileType {
   openai_api_key: string;
   stability_api_key: string;
   bria_api_key: string;
+  did_api_key: string;
   selectedAvatar: string;
   selectedTalkingPhoto: string;
   useCredits: boolean;
@@ -30,6 +31,7 @@ const defaultProfile: ProfileType = {
   openai_api_key: "",
   stability_api_key: "",
   bria_api_key: "",
+  did_api_key: "",
   selectedAvatar: "",
   selectedTalkingPhoto: "",
   useCredits: true
@@ -167,6 +169,7 @@ function createNewProfile(
     openai_api_key: "",
     stability_api_key: "",
     bria_api_key: "",
+    did_api_key: "",
     selectedAvatar: "",
     selectedTalkingPhoto: "",
     useCredits: true
